Move FitsImageSource code to cpp
Also, be more flexible about the header types supported from the FITS
file
