+++ SEFramework/SEFramework/FITS/FitsImageSource.h
@@ -24,25 +24,12 @@
 #ifndef _SEFRAMEWORK_IMAGE_FITSIMAGESOURCE_H_
 #define _SEFRAMEWORK_IMAGE_FITSIMAGESOURCE_H_
 
-#include <iostream>
-#include <fstream>
-#include <iomanip>
-#include <type_traits>
-
-#include <boost/lexical_cast.hpp>
-#include <boost/filesystem.hpp>
-#include <boost/filesystem/path.hpp>
-#include <boost/regex.hpp>
-#include <boost/algorithm/string.hpp>
-
-#include <fitsio.h>
-
-#include "ElementsKernel/Exception.h"
-
 #include "SEFramework/CoordinateSystem/CoordinateSystem.h"
 #include "SEFramework/Image/ImageSource.h"
-
 #include "SEFramework/FITS/FitsFileManager.h"
+
+#include <iostream>
+#include <boost/lexical_cast.hpp>
 
 
 namespace SourceXtractor {
@@ -60,185 +47,40 @@ public:
    * @param manager
    */
   FitsImageSource(const std::string &filename, int hdu_number = 0,
-      std::shared_ptr<FitsFileManager> manager = FitsFileManager::getInstance())
-    : m_filename(filename), m_manager(manager), m_hdu_number(hdu_number) {
-    int status = 0;
-    int bitpix, naxis;
-    long naxes[2] = {1,1};
-
-    auto fptr = m_manager->getFitsFile(filename);
-
-    if (m_hdu_number <= 0) {
-      int hdunum, hdutype;
-      fits_get_num_hdus(fptr, &hdunum, &status);
-      if (status != 0) {
-        throw Elements::Exception() << "Can't get the number of HDUs in the FITS file: " << filename;
-      }
-
-      m_hdu_number = 0;
-      do {
-        ++m_hdu_number;
-        switchHdu(fptr, m_hdu_number);
-        fits_get_hdu_type(fptr, &hdutype, &status);
-        if (status == 0 && hdutype == IMAGE_HDU) {
-          fits_get_img_param(fptr, 2, &bitpix, &naxis, naxes, &status);
-        }
-        if (status != 0) {
-          throw Elements::Exception() << "Can't get the type of the HDU " << m_hdu_number << " in the FITS file: "
-                                      << filename;
-        }
-      } while (m_hdu_number <=hdunum && !(hdutype == IMAGE_HDU && naxis == 2));
-      if (hdutype != IMAGE_HDU) {
-        throw Elements::Exception() << "Can't find 2D image in FITS file: " << filename;
-      }
-    }
-    else {
-      switchHdu(fptr, m_hdu_number);
-    }
-
-    fits_get_img_param(fptr, 2, &bitpix, &naxis, naxes, &status);
-    if (status != 0 || naxis != 2) {
-      throw Elements::Exception() << "Can't find 2D image in FITS file: " << filename << "[" << m_hdu_number << "]";
-    }
-
-    m_width = naxes[0];
-    m_height = naxes[1];
-
-    loadHeadFile();
-  }
+      std::shared_ptr<FitsFileManager> manager = FitsFileManager::getInstance());
 
   FitsImageSource(const std::string &filename, int width, int height,
                   const std::shared_ptr<CoordinateSystem> coord_system = nullptr,
-                  std::shared_ptr<FitsFileManager> manager = FitsFileManager::getInstance())
-    : m_filename(filename), m_manager(manager), m_hdu_number(1) {
-    m_width = width;
-    m_height = height;
+                  std::shared_ptr<FitsFileManager> manager = FitsFileManager::getInstance());
 
-    {
-      // CREATE NEW FITS FILE
-      int status = 0;
-      fitsfile* fptr = nullptr;
-      fits_create_file(&fptr, ("!"+filename).c_str(), &status);
-      if (status != 0) {
-        throw Elements::Exception() << "Can't create or overwrite FITS file: " << filename;
-      }
-      assert(fptr != nullptr);
+  virtual ~FitsImageSource() = default;
 
-      long naxes[2] = {width, height};
-      fits_create_img(fptr, getImageType(), 2, naxes, &status);
-
-      if (coord_system) {
-        auto headers = coord_system->getFitsHeaders();
-        for (auto& h : headers) {
-          std::ostringstream padded_key, serializer;
-          padded_key << std::setw(8) << std::left << h.first;
-
-          serializer << padded_key.str() << "= " << std::left << std::setw(70) << h.second;
-          auto str = serializer.str();
-
-          fits_update_card(fptr, padded_key.str().c_str(), str.c_str(), &status);
-          if (status != 0) {
-            char err_txt[31];
-            fits_get_errstatus(status, err_txt);
-            throw Elements::Exception() << "Couldn't write the WCS headers (" << err_txt << "): " << str;
-          }
-        }
-      }
-
-      std::vector<T> buffer(width);
-      for (int i = 0; i<height; i++) {
-        long first_pixel[2] = {1, i+1};
-        fits_write_pix(fptr, getDataType(), first_pixel, width, &buffer[0], &status);
-      }
-      fits_close_file(fptr, &status);
-
-      if (status != 0) {
-        throw Elements::Exception() << "Couldn't allocate space for new FITS file: " << filename;
-      }
-    }
-
-    auto fptr = m_manager->getFitsFile(filename, true);
-    switchHdu(fptr, m_hdu_number);
-  }
-
-  virtual ~FitsImageSource() {
-  }
-
-  virtual std::string getRepr() const override {
+  std::string getRepr() const override {
     return m_filename;
   }
 
-  virtual std::shared_ptr<ImageTile<T>> getImageTile(int x, int y, int width, int height) const override {
-    auto fptr = m_manager->getFitsFile(m_filename);
-    switchHdu(fptr, m_hdu_number);
-
-    auto tile = std::make_shared<ImageTile<T>>((const_cast<FitsImageSource*>(this))->shared_from_this(), x, y, width, height);
-
-    long first_pixel[2] = {x+1, y+1};
-    long last_pixel[2] = {x+width, y+height};
-    long increment[2] = {1, 1};
-    int status = 0;
-
-    auto image = tile->getImage();
-    fits_read_subset(fptr, getDataType(), first_pixel, last_pixel, increment,
-                 nullptr, &image->getData()[0], nullptr, &status);
-    if (status != 0) {
-      throw Elements::Exception() << "Error reading image tile from FITS file.";
-    }
-
-    return tile;
-  }
+  std::shared_ptr<ImageTile<T>> getImageTile(int x, int y, int width, int height) const override;
 
   /// Returns the width of the image in pixels
-  virtual int getWidth() const override {
+  int getWidth() const override {
     return m_width;
   }
 
   /// Returns the height of the image in pixels
-  virtual int getHeight() const  override {
+  int getHeight() const  override {
     return m_height;
   }
 
-  virtual void saveTile(ImageTile<T>& tile) override {
-    auto fptr = m_manager->getFitsFile(m_filename, true);
-    switchHdu(fptr, m_hdu_number);
-
-    auto image = tile.getImage();
-
-    int x = tile.getPosX();
-    int y = tile.getPosY();
-    int width = image->getWidth();
-    int height = image->getHeight();
-
-    long first_pixel[2] = {x+1, y+1};
-    long last_pixel[2] = {x+width, y+height};
-    int status = 0;
-
-    fits_write_subset(fptr, getDataType(), first_pixel, last_pixel, &image->getData()[0], &status);
-    if (status != 0) {
-      throw Elements::Exception() << "Error saving image tile to FITS file.";
-    }
-  }
+  void saveTile(ImageTile<T>& tile) override;
 
   template <typename TT>
   bool readFitsKeyword(const std::string& header_keyword, TT& out_value) {
-    if (m_header_overwrite.find(header_keyword) != m_header_overwrite.end()) {
-      out_value = boost::lexical_cast<TT>(m_header_overwrite.at(header_keyword));
+    auto i = m_header.find(header_keyword);
+    if (i != m_header.end()) {
+      out_value = boost::lexical_cast<TT>(i->second);
       return true;
-    } else {
-      auto fptr = m_manager->getFitsFile(m_filename);
-      switchHdu(fptr, m_hdu_number);
-
-      double keyword_value;
-      int status = 0;
-      fits_read_key(fptr, TDOUBLE, header_keyword.c_str(), &keyword_value, nullptr, &status);
-      if (status == 0) {
-        out_value = keyword_value;
-        return true;
-      } else {
-        return false;
-      }
     }
+    return false;
   }
 
   int getHDU() const {
@@ -247,60 +89,9 @@ public:
 
 private:
 
-  void switchHdu(fitsfile* fptr, int hdu_number) const {
-    int status = 0;
-    int hdu_type = 0;
+  void switchHdu(fitsfile* fptr, int hdu_number) const;
 
-    fits_movabs_hdu(fptr, hdu_number, &hdu_type, &status);
-
-    if (status != 0) {
-      throw Elements::Exception() << "Could not switch to HDU # " << hdu_number << " in file " << m_filename;
-    }
-    if (hdu_type != IMAGE_HDU) {
-      throw Elements::Exception() << "Trying to access non-image HDU in file " << m_filename;
-    }
-  }
-
-  void loadHeadFile() {
-    auto filename =  boost::filesystem::path(m_filename);
-    auto base_name = filename.stem();
-    base_name.replace_extension(".head");
-    auto head_filename = filename.parent_path() / base_name;
-
-    int current_hdu = 1;
-
-    if (boost::filesystem::exists(head_filename)) {
-      std::ifstream file;
-
-      // open the file and check
-      file.open(head_filename.native());
-      if (!file.good() || !file.is_open()){
-        throw Elements::Exception() << "Cannot load ascii header file: " << head_filename;
-      }
-
-      while (file.good()) {
-        std::string line;
-        std::getline(file, line);
-
-        line = boost::regex_replace(line, boost::regex("\\s*#.*"), std::string(""));
-        line = boost::regex_replace(line, boost::regex("\\s*$"), std::string(""));
-
-        if (line.size() == 0) {
-          continue;
-        }
-
-        if (boost::to_upper_copy(line) == "END") {
-          current_hdu++;
-        } else if (current_hdu == m_hdu_number){
-          boost::smatch sub_matches;
-          if (boost::regex_match(line, sub_matches, boost::regex("(.+)=(.+)")) && sub_matches.size()==3) {
-            auto keyword = boost::to_upper_copy(sub_matches[1].str());
-            m_header_overwrite[keyword] = sub_matches[2];
-          }
-        }
-      }
-    }
-  }
+  void loadHeadFile();
 
   int getDataType() const;
   int getImageType() const;
@@ -313,8 +104,7 @@ private:
 
   int m_hdu_number;
 
-
-  std::map<std::string, std::string> m_header_overwrite;
+  std::map<std::string, std::string> m_header;
 };
 
 }
